+++ spec/lib/karafka/setup/config_spec.rb
@@ -8,23 +8,10 @@ RSpec.describe Karafka::Setup::Config do
   end
 
   describe '#setup_components' do
-    before do
-      Karafka::Setup::Configurators::Base.descendants.each do |descendant_class|
-        config = double
-
-        expect(descendant_class)
-          .to receive(:new)
-          .with(config_class.config)
-          .and_return(config)
-          .at_least(:once)
-
-        expect(config)
-          .to receive(:setup)
-          .at_least(:once)
-      end
-    end
-
-    it 'expect to take descendants of BaseConfigurator and run setup on each' do
+    it 'expect to run setup for waterdrop' do
+      expect(Karafka::Setup::Configurators::WaterDrop)
+        .to receive(:setup).with(config_class.config)
+                           .at_least(:once)
       config_class.send :setup_components
     end
   end
@@ -67,6 +54,15 @@ RSpec.describe Karafka::Setup::Config do
         expect { config_class.send(:validate!) }
           .not_to raise_error
       end
+    end
+  end
+
+  describe '#after_init' do
+    it 'expect to call the after_init block' do
+      expect(config_class.config.internal.after_init)
+        .to receive(:call).with(config_class.config)
+
+      config_class.after_init
     end
   end
 end